split main.py to modules - models, db, beer router, utils
