+++ .gitignore
@@ -1,0 +1,6 @@
+__pycache__/
+archive/
+peewee_dir/
+routers/__pychache__/
+utils/__pychache__/
+config.json
